EDIT "Hall of Fame" on progress
in this session I have successfully passed an ID as parameter and with get request I fetched all the information from the controller to load it in the edit page, to show the current data. The edit page will be dynamic according to the fetched information. Up next I will complete the  Update operation.